Allow for negative longitudes and latitudes
Could you check if this one works for you? The regex should now be able to find positive and negative floats. On my phone (with lon, lat being both positive) it works.
I've also reverted the change regarding the accuracy measure. Since it can only be positive there is no need to adjust this..